# Misc optimizations for pyxl2