Please take a look at the the individual commits to figure out whats changed.
The switch from 4byte to a 2byte unicode type changed a lot of the unicode functions names so it maybe a good idea to regenerate the section ordering file but I was not sure with which perf sample I have to run it.
```
                                   upstream/master:        origin/pyxl2:
           django_template3.py             2.2s (4)             2.1s (2)  -1.8%
                 pyxl_bench.py             1.8s (4)             1.8s (2)  +1.3%
     sqlalchemy_imperative2.py             2.3s (4)             2.3s (2)  +0.7%
                pyxl_bench2.py             1.3s (4)             1.1s (2)  -19.5%
       django_template3_10x.py            14.1s (4)            13.6s (2)  -3.7%
             pyxl_bench_10x.py            13.9s (4)            14.2s (2)  +1.6%
 sqlalchemy_imperative2_10x.py            17.6s (4)            17.9s (2)  +1.4%
            pyxl_bench2_10x.py            11.8s (4)             9.1s (2)  -22.6%
                       geomean                 5.1s                 4.8s  -5.8%
```