@kmod 2015-10-21
Can we split out the unicode change?  Just double-checked that Dropbox's build is a wide build, so I'm not sure what kinds of issues would be caused by switching to a narrow build.  I wouldn't be surprised if it's fine to make the change (or if it's not) but I think it's a bit more complicated than the rest of the changes in the PR.

fyi you can run `make update_section_ordering` and it will do everything and update the section ordering file in place :)