@undingen 2015-10-22
I removed the unicode change
```
                                   upstream/master:        origin/pyxl2:
           django_template3.py             2.2s (4)             2.2s (2)  +0.1%
                 pyxl_bench.py             1.8s (4)             1.8s (2)  +0.7%
     sqlalchemy_imperative2.py             2.3s (4)             2.3s (2)  +1.2%
                pyxl_bench2.py             1.3s (4)             1.1s (2)  -14.9%
       django_template3_10x.py            14.2s (4)            14.3s (2)  +0.6%
             pyxl_bench_10x.py            14.3s (4)            13.9s (2)  -3.0%
 sqlalchemy_imperative2_10x.py            17.5s (4)            17.7s (2)  +1.1%
            pyxl_bench2_10x.py            12.2s (4)             9.6s (2)  -21.2%
                       geomean                 5.2s                 4.9s  -4.8%

```