@undingen 2015-10-27
I now using the much nicer PyArg_ParseSingle: https://github.com/undingen/pyston/commit/36ff3d23d1b9f6b85e53f0035f2a88292430ff99
and removed the gc untracke memory check: https://github.com/undingen/pyston/commit/82459763b0ff8e6d4ba6a3f3f6d0fba6fb1baa4b
Should be ready for merging